moving templates and renaming them
