+++ GCDS.NetTemplate.MVC.Sanity/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
 ﻿@{
-	Layout = "_Layout.Splash";
+	Layout = "GCDS.NetTemplate/_Layout.Splash";
 }
