+++ GCDS.NetTemplate.MVC.Sanity/Views/Home/Internal.cshtml
@@ -1,5 +1,5 @@
 ﻿@{
-    Layout = "_Layout.InternalApp";
+    Layout = "GCDS.NetTemplate/_Layout.InternalApp";
 }
 
 <h1>Hello!</h1>
