+++ GCDS.NetTemplate.MVC.Sanity/Views/_ViewStart.cshtml
@@ -1,3 +1,3 @@
 ﻿@{
-	Layout = "_Layout.Basic";
+	Layout = "GCDS.NetTemplate/_Layout.Basic";
 }
