+++ GCDS.NetTemplate.Razor.Sanity/Pages/Internal.cshtml
@@ -1,7 +1,7 @@
 ﻿@page
 @model InternalModel
 @{
-    Layout = "_Layout.InternalApp";
+    Layout = "GCDS.NetTemplate/_Layout.InternalApp";
     ViewData["Title"] = "Privacy Policy";
 }
 <h1>@ViewData["Title"]</h1>
