+++ GCDS.NetTemplate.Razor.Sanity/Pages/_ViewStart.cshtml
@@ -1,3 +1,3 @@
 ﻿@{
-    Layout = "_Layout.Basic";
+    Layout = "GCDS.NetTemplate/_Layout.Basic";
 }
